Turn scan/query/count extension functions into DynamoDbIndexMapper member functions
